+++ elasticsearch/config_test.go
@@ -1,0 +1,32 @@
+package elasticsearch
+
+import (
+	"context"
+	"errors"
+	"testing"
+)
+
+func TestUseCredentailsProvider(t *testing.T) {
+	t.Run("should return error if key is not set", func(t *testing.T) {
+		_, err := useCredentialsProvider(context.Background(), "", "", false)
+		if !errors.Is(err, errCredentialProviderKeyNotSet) {
+			t.Errorf("expected error to be errCredentialProviderKeyNotSet, got %v", err)
+		}
+	})
+
+	t.Run("key is set", func(t *testing.T) {
+		t.Run("should return error if mechanism is not set", func(t *testing.T) {
+			_, err := useCredentialsProvider(context.Background(), "key", "", false)
+			if !errors.Is(err, errCredentialProviderMechanismNotSet) {
+				t.Errorf("expected error to be errCredentialProviderMechanismNotSet, got %v", err)
+			}
+		})
+
+		t.Run("should return error if mechanism is invalid", func(t *testing.T) {
+			_, err := useCredentialsProvider(context.Background(), "key", "invalid", false)
+			if !errors.Is(err, errCredentialProviderMechanismInvalid) {
+				t.Errorf("expected error to be errCredentialProviderMechanismInvalid, got %v", err)
+			}
+		})
+	})
+}